requirements.txt
wheel
pytorch
torchvision
torchaudio
numpy
matplotlib
tqdm
pylint
torchsummary
webdataset